Moved entities into single file
