+++ despatchbay/despatchbay_sdk.py
@@ -1,22 +1,10 @@
 from suds.client import Client
 from suds.transport.http import HttpAuthenticated
 import suds
-from exception import AuthorizationException,\
+from despatchbay.exceptions import AuthorizationException,\
     ApiException, ConnectionException, RateLimitException
-from despatchbay.entities.parcel import Parcel
-from despatchbay.entities.address import Address
-from despatchbay.entities.recipient import Recipient
-from despatchbay.entities.sender import Sender
-from despatchbay.entities.shipment_request import ShipmentRequest
-from despatchbay.entities.account import Account
-from despatchbay.entities.account_balance import AccountBalance
-from despatchbay.entities.address_key import AddressKey
-from despatchbay.entities.service import Service
-from despatchbay.entities.collection import Collection
-from despatchbay.entities.shipment_return import ShipmentReturn
-from despatchbay.entities.payment_method import PaymentMethod
-from despatchbay.entities.automatic_topup_settings import AutomaticTopupSettings
-from despatchbay.entities.collection_date import CollectionDate
+from despatchbay.entities import Parcel, Address, Recipient, Sender, ShipmentRequest, Account, AccountBalance,\
+    AddressKey, Service, Collection, ShipmentReturn, PaymentMethod, AutomaticTopupSettings, CollectionDate
 from despatchbay.documents_client import DocumentsClient
