+++ despatchbay/exceptions.py
