+++ setup.py
@@ -1,0 +1,11 @@
+from setuptools import setup
+
+setup(name='despatchbay',
+      version='0.1',
+      description='Despatchbay SDK',
+      url='',
+      author='',
+      author_email='',
+      license='MIT',
+      packages=['despatchbay'],
+      zip_safe=False)
